changes para erros ao não encontrar o objeto

+++ SPRING-REST-API/forum/forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -1,12 +1,13 @@
 package br.com.alura.forum.controller;
 
+import br.com.alura.forum.dto.DetalhesTopicoDTO;
 import br.com.alura.forum.dto.TopicoDTO;
 import br.com.alura.forum.form.AtualizacaoTopico;
 import br.com.alura.forum.form.TopicoForm;
 import br.com.alura.forum.modelo.Topico;
 import br.com.alura.forum.repository.CursoRepository;
 import br.com.alura.forum.repository.TopicoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -15,14 +16,14 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
+@AllArgsConstructor
 @RestController
 @RequestMapping(value = "/topicos")
 public class TopicosController {
 
-    @Autowired
     private TopicoRepository topicoRepository;
-    @Autowired
     private CursoRepository cursoRepository;
 
 
@@ -43,25 +44,38 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public TopicoDTO detalhar(@PathVariable Long id) {
-        Topico topico = topicoRepository.getOne(id);
-        return new TopicoDTO(topico);
+    public ResponseEntity<DetalhesTopicoDTO> detalhar(@PathVariable Long id) {
+        Optional<Topico> topico = topicoRepository.findById(id);
+        if (topico.isPresent()) {
+            return ResponseEntity.ok(new DetalhesTopicoDTO(topico.get()));
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @Transactional
     @PutMapping("/{id}")
     public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id,
                                                @RequestBody @Valid AtualizacaoTopico form) {
-        Topico topico = form.atualizar(id, topicoRepository);
-        return ResponseEntity.ok(new TopicoDTO(topico));
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if (optional.isPresent()) {
+            Topico topico = form.atualizar(id, topicoRepository);
+
+            return ResponseEntity.ok(new TopicoDTO(topico));
+
+        }
+        return ResponseEntity.notFound().build();
 
     }
 
     @Transactional
     @DeleteMapping("/{id}")
     public ResponseEntity<?> remover(@PathVariable Long id) {
-        topicoRepository.deleteById(id);
-        return ResponseEntity.noContent().build();
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if (optional.isPresent()) {
+            topicoRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
